requirements/base.txt
fastapi
uvicorn[standard]

sqlalchemy
psycopg[binary]

alembic
python-dotenv

pydantic[email]

python-jose[cryptography]
passlib==1.7.4
bcrypt==3.2.2